pyproject.toml
[project]
name = "storefront"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-debug-toolbar>=4.4.6",
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "psycopg2-binary>=2.9.10",
]
